Split some multi-use functions off into a separate file
-declutters the main files

-removes the need for a circular dependency
